add danish agreement to pp dto and refactor contract created event
