+++ src/main/java/com/hedvig/notificationService/serviceIntegration/productPricing/client/DanishAccidentLineOfBusiness.kt
@@ -1,0 +1,6 @@
+package com.hedvig.notificationService.serviceIntegration.productPricing.client
+
+enum class DanishAccidentLineOfBusiness {
+    REGULAR,
+    STUDENT
+}
