+++ src/main/java/com/hedvig/notificationService/serviceIntegration/productPricing/client/DanishHomeContentLineOfBusiness.kt
@@ -1,0 +1,8 @@
+package com.hedvig.notificationService.serviceIntegration.productPricing.client
+
+enum class DanishHomeContentLineOfBusiness {
+    RENT,
+    OWN,
+    STUDENT_RENT,
+    STUDENT_OWN
+}
