+++ src/main/java/com/hedvig/notificationService/serviceIntegration/productPricing/client/DanishTravelLineOfBusiness.kt
@@ -1,0 +1,6 @@
+package com.hedvig.notificationService.serviceIntegration.productPricing.client
+
+enum class DanishTravelLineOfBusiness {
+    REGULAR,
+    STUDENT
+}
